Extract sub-section "Addenda"
OK, so you go through a bunch of steps and come to these instructions:

> There should now be a CiviCRM link in your Drupal menu.  Click that link and
> the CiviCRM Menu, Shortcuts, Search and New Individual Blocks should appear.
> You can now explore CiviCRM end-user features and begin configuring CiviCRM
> for your site/organization needs.

This should be the end of the installation guide. But, if you skim the TOC,
it keeps on going!

I *think* what's going on here is that people wanted to add some sections
which are sort of independent of the main flow of installation.

These probably should be moved/referenced in some other way, but I don't
really want to think too hard.  This commit just mitigates the confusion
a bit by setting the subsequent sections (re:MySQL encryption and troubleshooting)
under the heading "Addenda".